tercer integradora

+++ src/routes/apiProducts.router.js
@@ -13,7 +13,8 @@ const router = Router();
 router.get('/', auth, isVerified, async (req, res) => {
   const limit = +req.query.limit || 8;
   const page = +req.query.page || 1;
-  const { query = null, sort = null } = req.query;
+  let query = req.query.query || null;
+  let sort = req.query.sort || null;
   
   if (typeof limit !== 'number' || typeof page !== 'number') {
     req.logger.warning(`${req.method} ${req.path} - limit and page must be numbers`)
@@ -23,14 +24,17 @@ router.get('/', auth, isVerified, async (req, res) => {
     req.logger.warning(`${req.method} ${req.path} - limit and page must be greater than 0`)
     return res.status(400).send({status: 'error', message: 'limit and page must be greater than 0'})
   }
-  if (query) {
-    query = JSON.parse(query);
-  }
-  if (sort) {
-    sort = JSON.parse(sort)
-  }
 
+  console.log(query, sort)
+  
   try {
+    if (query) {
+      query = JSON.parse(query);
+    }
+    if (sort) {
+      sort = JSON.parse(sort)
+    }
+
     const products = await productService.getProducts(limit, page, query, sort);
     res.status(200).send({status: 'success', message: 'productos encontrados', products})
   } catch (error) {
@@ -51,8 +55,21 @@ router.get('/:productId', auth, isVerified, async (req, res) => {
   }
 })
 
-router.post('/', auth, isVerified, isAdmin, async (req, res) => {
-  const product = req.body.product;
+router.post('/', auth, isVerified, async (req, res) => {
+  const product = {
+    title: req.body.product.title,
+    description: req.body.product.description,
+    code: req.body.product.code,
+    price: parseInt(req.body.product.price),
+    stock: parseInt(req.body.product.stock),
+    category: req.body.product.category,
+    thumbnails: req.body.product.thumbnails,
+  }
+
+  if (req.session.user.role == "usuario") {
+    req.logger.warning(`${req.method} ${req.path} - no tiene permisos para agregar productos`)
+    return res.status(403).send({status: 'error', message: 'no tiene permisos para agregar productos'})
+  }
   
   try {
     if (!product.title || !product.description || !product.code || !product.price || !product.stock || !product.category || !product.thumbnails) {
@@ -65,20 +82,30 @@ router.post('/', auth, isVerified, isAdmin, async (req, res) => {
       });
     }
 
-    if (typeof product.title !== 'string' || typeof product.description !== 'string' || typeof product.code !== 'number' || typeof product.price !== 'number' || product.price <= 0 || typeof product.stock !== 'number' || product.stock < 0 || typeof product.category !== 'string' || !Array.isArray(product.thumbnails)) {
+    if (typeof product.title !== 'string' || typeof product.description !== 'string' || typeof product.price !== 'number' || product.price <= 0 || typeof product.stock !== 'number' || product.stock < 0 || typeof product.category !== 'string') {
       req.logger.warning(`${req.method} ${req.path} - One or more fields have the wrong type`)
       CustomError.createError({
         name: "Product Error",
-        cause: generateProductsErrorInfo(errorCodes.INVALID_TYPES_ERROR, req.body.product),
+        cause: generateProductsErrorInfo(errorCodes.INVALID_TYPES_ERROR, product),
         message: "One or more fields have the wrong type",
         code: errorCodes.INVALID_TYPES_ERROR,
       });
     }
 
-    const newProduct = await productService.addProduct(product)
+    const newProduct = await productService.addProduct({
+      title: product.title,
+      description: product.description,
+      code: product.code,
+      price: product.price,
+      stock: product.stock,
+      category: product.category,
+      thumbnails: product.thumbnails,
+      owner: req.session.user._id
+    })
+
     res.status(201).send({status:'success', message:'producto agregado', newProduct})
   } catch (error){
-    req.logger.error(`${req.method} ${req.path} - ${error.message}`)
+    req.logger.error(`${req.method} ${req.path} - ${error}`)
     res.status(400).send({status:'error', error: error})
   }
 })
@@ -114,12 +141,30 @@ router.put('/:productId', auth, isVerified, isAdmin, async (req, res) => {
   }
 })
 
-router.delete('/:productId', auth, isVerified, isAdmin, async (req, res) => {
+router.delete('/:productId', auth, isVerified, async (req, res) => {
   const productId = req.params.productId;
   
   try {
-    const product = await productService.deleteProduct(productId);
-    res.status(200).send({status:'success', message:'producto eliminado', product})
+    const product = await productService.getProductById(productId);
+
+    if (!product) {
+      req.logger.warning(`${req.method} ${req.path} - producto no encontrado`)
+      return res.status(404).send({status: 'error', message: 'producto no encontrado'})
+    }
+
+    if (req.session.user.role === 'admin') {
+      await productService.deleteProduct(productId);
+      return res.status(200).send({status:'success', message:'producto eliminado'})
+    }
+
+    if (product.owner.toString() !== req.session.user._id.toString()) {
+      req.logger.warning(`${req.method} ${req.path} - no tiene permisos para eliminar este producto`)
+      return res.status(403).send({status: 'error', message: 'no tiene permisos para eliminar este producto'})
+    }
+
+    await productService.deleteProduct(productId);
+    res.status(200).send({status:'success', message:'producto eliminado'})
+
   } catch (error){
     req.logger.error(`${req.method} ${req.path} - ${error.message}`)
     res.status(400).send({status:'error', message: error.message})
